+++ server/src/rooms/GameRoom.ts
@@ -1,9 +1,10 @@
 import { Room, Client } from "colyseus";
 import { environment } from "../environment";
 import { createLogger } from "../logger";
-import { Vector } from "../math/Vector";
 import { DirectionMessage } from "./messages/DirectionMessage";
-import { Body, GameState, Player } from "./schema/GameState";
+import { BodySchema, GameState, Player } from "./schema/GameState";
+import { runGameSimulation } from "./Simulation";
+import { World, Bodies, Runner, Engine, Body, Vector } from "matter-js";
 const log = createLogger("gameroom");
 
 export class ServerBody {
@@ -12,8 +13,12 @@ export class ServerBody {
 
 export class MyRoom extends Room<GameState> {
   lastBodyId = 1;
+  engine!: Engine;
+  matterMap = new Map<BodySchema, Body>();
+
   onCreate(options: any) {
     this.setState(new GameState());
+    this.matterMap = new Map<BodySchema, Body>();
 
     this.onMessage<DirectionMessage>("direction", (client, message) => {
       log("direction message", message, "from", client.sessionId);
@@ -27,7 +32,7 @@ export class MyRoom extends Room<GameState> {
         if (message.right) moveX++;
         if (moveX !== 0 || moveY !== 0) {
           log("move direction %j", { moveX, moveY });
-          const { x, y } = Vector.of(moveX, moveY).normalize();
+          const { x, y } = Vector.normalise(Vector.create(moveX, moveY));
           log("after direction %j", { x, y });
           body.moveDirection.x = x;
           body.moveDirection.y = y;
@@ -42,33 +47,62 @@ export class MyRoom extends Room<GameState> {
       (millis) => this.update(millis),
       environment.SIMULATION_INTERVAL
     );
+    this.createPhysics();
+  }
+  private createPhysics() {
+    this.engine = Engine.create({ gravity: { x: 0, y: 0 } });
   }
 
   update(millis: number) {
-    this.state.players.forEach((player) => {
-      const body = this.state.bodies.get(player.bodyId);
-      if (body) {
-        const speed = body.speed;
-        body.position.x += body.moveDirection.x * millis * speed;
-        body.position.y += body.moveDirection.y * millis * speed;
-      }
+    // update body forces / directions
+    this.state.bodies.forEach((body) => {
+      const matterBody = this.matterMap.get(body);
+      if (!matterBody) return;
+      const speed = body.speed;
+      const force = Vector.mult(body.moveDirection, speed * millis * 0.002);
+      Body.applyForce(matterBody, body.position, force);
+      matterBody.friction = 0.9;
+      matterBody.frictionAir = 0.9;
+      matterBody.frictionStatic = 0.9;
     });
+    // run physics simulation
+    Engine.update(this.engine, millis);
+    // sync positions
+    for (const [bodySchema, matterBody] of this.matterMap) {
+      bodySchema.position.x = matterBody.position.x;
+      bodySchema.position.y = matterBody.position.y;
+    }
   }
 
   onJoin(client: Client, options: any) {
     log(client.sessionId, "joined!");
     const player = new Player(client.sessionId, "<player name>");
-    const body = new Body(`${this.lastBodyId++}`);
+    const body = new BodySchema(`${this.lastBodyId++}`);
+    body.radius = 16;
     player.bodyId = body.id;
     this.state.players.set(player.id, player);
     this.state.bodies.set(body.id, body);
+    const matterBody = Bodies.circle(
+      body.position.x,
+      body.position.y,
+      body.radius
+    );
+    World.add(this.engine.world, matterBody);
+    this.matterMap.set(body, matterBody);
   }
 
   onLeave(client: Client, consented: boolean) {
     log(client.sessionId, "left!");
     const { player, body } = this.state.findPlayerAndBody(client.sessionId);
     if (player) this.state.players.delete(player.id);
-    if (body) this.state.bodies.delete(body.id);
+    if (body) {
+      this.state.bodies.delete(body.id);
+      const matterBody = this.matterMap.get(body);
+      if (matterBody) {
+        World.remove(this.engine.world, matterBody);
+        this.matterMap.delete(body);
+      }
+    }
   }
 
   onDispose() {