+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ## 0.1.0
 
+- integrate matter js as physics engine
 - client can control player bodies
 - player synchronizes across clients